Setup project layout, docs, htcondor subpackage
The first commit containing real code in the form of the htcondor
subpackage, it also sets up the directory layout and of the project and
adds the docs directory for automatic documentation builds by Read the Docs.

So far these have been forced pushes upstream, but once docs build
failures and other cosmetic issues are ironed out, I'll go back to
proper pushes. ;]