Rerelease version 0.1.0.a1
The previous commmit did not resolve the issue with missing data files.
This StackOverflow solution https://stackoverflow.com/a/14159430 helped
by explaining that for data files to be included in sdist, they must be
listed in MANIFEST.in, whereas include_package_data is needed for them
to be included in the bdist wheel.