Rewrite/refactor of HTCondorized, cosmetic changes
While the previous version of HTCondorized worked fine on lxplus, it
broke on cmslpc for a few reasons:

* The HTCondor version is currently 8.4 on cmslpc and 8.5 on lxplus.
This bump in minor version saw the addition of an ALL_NODES option for
DAGMAN applications, which wasn't available in previous versions. This
option was being used to specify the retrying of all nodes on a single
line. For the sake of compatibility, there is now an explicit RETRY
command for each node, since this pattern is recognized across versions.
* The "getenv = True" command does not work on cmslpc. This option works
fine on lxplus and is used for transparent copying of the submission
environment variables, specifically those pertaining to the Python
virtual environment.

  After sending some test submissions at cmslpc, it is evident that the
submission environment is being copied, but something is prepending to
PATH in the execution nvironment such that the system Python interpreter
/usr/bin/python is always found before the virtualenv Python.

  The commonly recommended solution is to tar your working area and
submit it along with the rest of the input files. However, that won't
work with a Python virtualenv because the paths are actually **hardcoded**
inside the venv/bin/activate script. This might not be a concern if the
filesystem is shared between the submission and execution environments,
but it may come back to bite us when this is not the case.

  The solution I used was to bootstrap the execution environment, taking
cues from this [cmslpc example](http://uscms.org/uscms_at_work/physics/computing/setup/batch_systems.shtml#condor_good_ex3).
This approach is more true to the spirit of keeping the submission and
execution environments indepdent, and only assumes that the execution
environment has access the same remote resources as the submission
environment.

  This approach works on both cmslpc and lxplus, so the "getenv" command
has been removed from the node submit description template, along with
some minor edits, and a new template introduced for a worker.sh script
which does the bootstrapping.

Implementing these changes to HTCondorized also gave me time to see a
way to refactor the submit method to better separate concerns.

The remaining cosmetic changes were some capitalization of constants
and removing of spurious whitespace.