Minor changes to HTCondorized behaviour
The timestamp should now correctly be timezone aware. Furthermore, the
deployment of CMSSW and the Python environment have been separated in
the worker.sh script to more clearly show the user the purpose of the
functions. This will also provide a point of reference for manual edits
to be made in order for the script to run in environments where neither
pip nor virtualenv have been installed (to be explained in the docs,
which are coming).

There's one cosmetic change to the setup.py file, and that is explicitly
stating the string formatting index for the version keyword argument in
the call to setup.