Refactor with metaclass, modify 3rd party licenses
The use of a metaclass was in the end misguided because it messes with
the method resolution order in unintuitive ways. While there may be use
cases for metaclasses, this like many others is not one of them.

And switching back to regular old classes means I can use the
cached_property decorator instead of rolling my own hack at a cached
class property decorator.