Merge conf subpackage into core
After further development, having a conf subpackage dedicated to
defining classes that end users customize was unnecessary.

Just as well, having DatasetBase and MonteCarloBase was equally
unnecessary. They have been eschewed in favor of a single Dataset class
which implements a property named `datatype`. The datatype is inferred
from the assignment, or lack thereof, of a cross section value.