Fix: Error in `python': double free or corruption
Iterating over a dataset and manipulating the events would sometimes
crash with this error. Based on the stacktrace, it appears that PyROOT
didn't like the way eventlists were named by hashing the selection.

Performing the same selection on different Events instances would create
eventlists with the same name, and ROOT's penchant for global state
might be the cause of the crash?

Moving to a universally unique identifier (UUID) appears to have solved
the issue.